fix: apply cargo fmt workspace-wide and resolve clippy -D warnings
Phase 0 ("stop the bleeding") of the productionizing plan:

- cargo fmt across the entire 1.53M-line workspace (~1942 files were
  unformatted; this is the mechanical part of that diff).
- Fix the two mcx::mut_from_ref deny-level soundness errors (aset_mut/
  bump_mut returning &mut T with an aliasable lifetime from &UnsafeCell<T>)
  by returning *mut T instead and wrapping call sites in unsafe blocks.
  This was blocking clippy across virtually the whole tree: of 848
  workspace members only 28 were reachable before this fix.
- Resolve everything clippy -D warnings turned up once the rest of the
  tree became reachable: raw-pointer-arg safety docs, needless range
  loops, unnecessary unsafe blocks, dead code, type-complexity aliases,
  non-idiomatic boolean/control-flow patterns, and a handful of
  crate-scoped allows for lints that would misfire on intentional C
  fidelity (verbatim-ported names, mechanically generated snowball
  stemmer code, cascading phonetic-algorithm branches, etc.) — each
  documented at its site.
- Root-cause and fix the flaky execmain start/park/rearm test: its
  QueryDescHandle tap compared handle values across threads, but the
  underlying slot tables are thread_local!, so handles only compare
  uniquely within a single thread. The tap now also gates on the
  originating thread id.

`cargo clippy --workspace --all-targets -- -D warnings` is not yet fully
green — clippy stops at the first broken crate per dependency subtree, so
each fix here reveals the next crate's warnings on the following run. This
commit captures the fixes applied so far; the loop continues.

+++ crates/backend/access/heap/heapam/src/hio.rs
@@ -270,7 +270,7 @@ pub fn RelationGetBufferForTuple<'mcx>(
     // at use time in dml.rs heap_multi_insert instead, which owns the FROZEN
     // all-visible marking + visibilitymap_set (heapam.c:2460-2654).
 
-    let save_free_space = relation.get_target_page_free_space(HEAP_DEFAULT_FILLFACTOR) as usize;
+    let save_free_space = relation.get_target_page_free_space(HEAP_DEFAULT_FILLFACTOR);
     let nearly_empty_free_space =
         MaxHeapTupleSize - (MaxHeapTuplesPerPage / 8 * core::mem::size_of::<ItemIdData>());
     let target_free_space = if len + save_free_space > nearly_empty_free_space {
